xpath 1.13-7 from libxml-xpath-perl
The xpath script included with Mac OS X is basically unusable.
There is no -q flag to remove the prefix '-- NODE --' and attempts to pipe
the stderr to /dev/null removes the new line separating the actual
nodes found.

This version of the script actually had a man page with useful
information, unlike the one out of the box.